Update with private annotations and renaming
